+++ NEWS.md
@@ -4,7 +4,7 @@
   - dutility function no longer casts warnings when handling formulas
   - Efficient estimation of recurrent events mean: recurrentMarginalAIPCW 
   - Average treatment effect for competing risks and binary data: logitATE, binregATE
-  - Recurrent events regression with IPCW adjustment : recreg
+  - Recurrent events regression with IPCW adjustment (Ghosh-Lin model) : recreg
 
 # mets 1.2.8.1
   - Maintenance release